added iso_env_fortran

+++ src/Modules/Fobjtype.f90
@@ -1,6 +1,7 @@
 module fobjtype
-    
-    logical::fobj
-    integer::num_it_conv,num_it_opt
+    use iso_fortran_env
+
+    logical :: fobj
+    integer(kind=int32) :: num_it_conv,num_it_opt
     
 endmodule fobjtype
